Rename machinecontroller to machine-controller.
This makes the naming more consistent with other components.
