+++ cluster-api/machine-controller/.gitignore
@@ -1,0 +1,1 @@
+machinecontroller
