+++ cluster-api/machine-controller/Dockerfile
@@ -1,0 +1,5 @@
+FROM busybox
+
+ADD machine-controller /bin
+
+ENTRYPOINT ["/bin/machine-controller"]
