--- cluster-api/machinecontroller/.gitignore
@@ -1,1 +1,0 @@
-machinecontroller
