--- cluster-api/machinecontroller/Dockerfile
@@ -1,5 +1,0 @@
-FROM busybox
-
-ADD machinecontroller /bin
-
-ENTRYPOINT ["/bin/machinecontroller"]
